Time: 0 ms (100.00%) | Memory: 8.7 MB (14.59%) - LeetSync

+++ 69-sqrtx/sqrtx.cpp
@@ -4,8 +4,8 @@ public:
           int l=1,r=x;
           while (l<=r)
           {
-            long long mid= l+(r-l)/2;
-            long long val=(mid*mid);
+            long  mid= l+(r-l)/2;
+            long  val=(mid*mid);
             if(val<=x)
             {
                 l=mid+1;
